requirements.txt
sqlalchemy>=2.0
psycopg[binary]>=3.1
pydantic>=2.7
python-dotenv>=1.0